requirements.txt
# Core dependencies for WAT Framework

# Environment management
python-dotenv==1.0.0

# HTTP requests and web scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
validators==0.22.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
gspread==6.0.0

# Data processing
pandas==2.2.0
numpy>=1.26.0
openpyxl==3.1.2

# ML / Orders Estimator
lightgbm>=4.0.0
scikit-learn>=1.4.0
scipy>=1.12.0

# AI/ML APIs
openai==1.10.0
anthropic==0.18.1

# Browser automation
playwright==1.40.0

# Utilities
pyyaml==6.0.1
